Version 1 of script
pianoboxette can be used to push files to dropbox based on the oauth
key set in the environment variable.

file and url args are supported although the url just downloads the file
locally temporarily. I was too lazy to figure out a clean way to just
read the bytes into mem and then use it in the upload step. If I ever
revisit this I can make it faster that way since it will avoid the disk
write/read